backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.5.0
python-decouple==3.8
drf-spectacular==0.29.0